+++ egg/zoo/emcom_as_ssl/utils.py
@@ -17,12 +17,6 @@ def get_data_opts(parser):
         help="Dataset location",
     )
     group.add_argument(
-        "--validation_dataset_dir",
-        type=str,
-        default="/datasets01/imagenet_full_size/061417/val",
-        help="Dataset location",
-    )
-    group.add_argument(
         "--image_size",
         type=int,
         default=224,
